+++ src/call_frame.rs
@@ -18,7 +18,7 @@ pub struct CallFrame {
 pub struct CodePage(Vec<U256>);
 
 impl CodePage {
-    pub fn get(&self, idx: u16) -> U256 {
+    pub fn get(&self, idx: usize) -> U256 {
         // NOTE: the spec mandates reads past the end of the program return any value that decodes
         // as an `invalid` instruction. 0u256 fits the bill because its decoded variant is 0 which
         // in turn is **the** invalid opcode.
